# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Swagger Bootcamp - OpenAPI 3.0
  description: Contrato de Banco Bootcamp
  version: 1.0.0
servers:
  - url: nttdata/
...
paths:
...
  /company:
    get:
      tags:
        - Company
      summary: Find All Customers Company
      description: Find All Customers Company
      operationId: findAllCompany
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyResponseDto'
    put:
      tags:
        - Company
      summary: Update an existing customer company
      description: Update an existing customer company by Id
      operationId: updateCompany
      requestBody:
        description: Update an existent customer company in the bank
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyRequestDto'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyResponseDto'
    post:
      tags:
        - Company
      summary: Add a new customer company to the bank
      description: Add a new customer company to the bank
      operationId: createCompany
      requestBody:
        description: Create a new customer company in the bank
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyRequestDto'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyResponseDto'
  /company/{companyId}:
    get:
      tags:
        - Company
      summary: Find Customers Company by id
      description: Find Customers Company by id
      operationId: findCompanyById
      parameters:
        - name: companyId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyResponseDto'
...
  /creditCard:
    get:
      tags:
        - CreditCard
      summary: Find All CreditCard
      description: Find All CreditCard
      operationId: findAllCreditCard
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCardResponseDto'
    put:
      tags:
        - CreditCard
      summary: Update an existing CreditCard
      description: Update an existing CreditCard by Id
      operationId: updateCreditCard
      requestBody:
        description: Update an existent CreditCard in the bank
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditCardRequestDto'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCardResponseDto'
  /creditCard/person:
    post:
      tags:
        - CreditCard
      summary: Add a new creditCard person to the bank
      description: Add a new creditCard person to the bank
      operationId: createCreditCardPerson
      requestBody:
        description: Create a new CreditCard person in the bank
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditCardRequestDto'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCardResponseDto'
  /creditCard/company:
    post:
      tags:
        - CreditCard
      summary: Add a new CreditCard company to the bank
      description: Add a new CreditCard company to the bank
      operationId: createCreditCardCompany
      requestBody:
        description: Create a new CreditCard company in the bank
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditCardRequestDto'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCardResponseDto'
  /creditCard/{creditCardId}:
    get:
      tags:
        - CreditCard
      summary: Find CreditCard by id
      description: Find CreditCard by id
      operationId: findCreditCardById
      parameters:
        - name: creditCardId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCardResponseDto'
  /creditCard/customer/{customerId}:
    get:
      tags:
        - CreditCard
      summary: Find CreditCard by customerId
      description: Find CreditCard by customerId
      operationId: findCreditCardByCustomerId
      parameters:
        - name: customerId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCardResponseDto'
  /creditCard/pay:
    post:
      tags:
        - CreditCard
      summary: Pay a creditCard person or company to the bank
      description: Pay a creditCard person or company to the bank
      operationId: payCreditCard
      requestBody:
        description: Pay a CreditCard person or company in the bank
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditCardPayRequestDto'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCardPayResponseDto'
  /creditCard/consume:
    post:
      tags:
        - CreditCard
      summary: Consume credit card balance
      description: consume credit card balance
      operationId: consumeCreditCard
      requestBody:
        description: consume credit card balance
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditCardConsumeRequestDto'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCardConsumeResponseDto'
  /account:
    get:
      tags:
        - Account
      summary: Find All Account
      description: Find All Account
      operationId: findAllAccount
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponseDto'
    put:
      tags:
        - Account
      summary: Update an existing Account
      description: Update an existing Account by Id
      operationId: updateAccount
      requestBody:
        description: Update an existent Account in the bank
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRequestDto'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponseDto'
  /account/person:
    post:
      tags:
        - Account
      summary: Add a new account person to the bank
      description: Add a new account person to the bank
      operationId: createAccountPerson
      requestBody:
        description: Create a new account person in the bank
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRequestDto'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponseDto'
  /account/company:
    post:
      tags:
        - Account
      summary: Add a new account company to the bank
      description: Add a new account company to the bank
      operationId: createAccountCompany
      requestBody:
        description: Create a new account company in the bank
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRequestDto'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponseDto'
  /account/{accountId}:
    get:
      tags:
        - Account
      summary: Find Account by id
      description: Find Account by id
      operationId: findAccountById
      parameters:
        - name: accountId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponseDto'
  /account/typeAccount/{typeAccount}:
    get:
      tags:
        - Account
      summary: Find Account by type account
      description: Find Account by type account
      operationId: findAccountByTypeAccount
      parameters:
        - in: path
          name: typeAccount
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponseDto'
  /account/customer/{customerId}:
    get:
      tags:
        - Account
      summary: Find Account by customerId
      description: Find Account by customerId
      operationId: findAccountByCustomerId
      parameters:
        - name: customerId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponseDto'
  /account/deposit:
    post:
      tags:
        - Account
      summary: Deposit into account
      description: Deposit into account
      operationId: depositAccount
      requestBody:
        description: Deposit into account
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountDepositRequestDto'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDepositResponseDto'
  /account/withdraw:
    post:
      tags:
        - Account
      summary: Withdraw from account
      description: Withdraw from account
      operationId: withdrawAccount
      requestBody:
        description: Withdraw from account
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountWithdrawRequestDto'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountWithdrawResponseDto'
components:
  schemas:
...
    Company:
      required:
        - businessName
        - email
        - holders
        - id
        - ruc
        - telephone
      type: object
      properties:
        id:
          type: string
          example: 6413619bf92d28ec5a2e104e
        businessName:
          type: string
          example: Umbral
        ruc:
          type: string
          example: "10256987580"
        email:
          type: string
          example: contact@umbral.com
        telephone:
          type: string
          example: "259864"
        profile:
          type: string
          example: ESTANDAR
        holders:
          type: array
          items:
            $ref: '#/components/schemas/AuthorizedDto'
        signatories:
          type: array
          items:
            $ref: '#/components/schemas/AuthorizedDto'
    AuthorizedDto:
      required:
        - dni
        - email
        - lastName
        - name
        - telephone
      type: object
      properties:
        name:
          type: string
          example: Antonio
        lastName:
          type: string
          example: Silva
        dni:
          type: string
          example: "76893214"
        email:
          type: string
          example: antsi@correo.com
        telephone:
          type: string
          example: "945856325"
    CompanyResponseDto:
      required:
        - company
        - statusDto
      type: object
      properties:
        statusDto:
          $ref: '#/components/schemas/MessageStatus'
        company:
          type: array
          items:
            $ref: '#/components/schemas/Company'
    MessageStatus:
      required:
        - message
        - statusCode
      type: object
      properties:
        statusCode:
          type: string
          example: "0"
        message:
          type: string
          example: Operacion exitosa.
    CompanyRequestDto:
      required:
        - businessName
        - email
        - holders
        - ruc
        - telephone
      type: object
      properties:
        id:
          type: string
          example: 6413619bf92d28ec5a2e104e
        businessName:
          type: string
          example: Umbral
        ruc:
          type: string
          example: "10256987580"
        email:
          type: string
          example: contact@umbral.com
        telephone:
          type: string
          example: "259864"
        holders:
          type: array
          items:
            $ref: '#/components/schemas/AuthorizedDto'
        signatories:
          type: array
          items:
            $ref: '#/components/schemas/AuthorizedDto'
...
    CreditCard:
      required:
        - createDate
        - creditCardAccount
        - creditCardId
        - creditLine
        - customerId
        - feeMonth
        - id
        - lineAvailable
        - minimumAmount
        - payDay
        - settlementAmount
      type: object
      properties:
        id:
          type: string
          example: 6413619bf92d28ec5a2e104e
        customerId:
          type: string
          example: 6413619bf92d28ec5a2e104e
        creditLine:
          type: number
          format: double
          example: 1000
        creditCardAccount:
          type: string
          example: 6413619bf92d28ec5a2e104e
        creditCardId:
          type: string
          example: 1111-2222-3333-4444
        lineAvailable:
          type: number
          format: double
          example: 1000
        feeMonth:
          type: number
          format: double
          example: 0
        minimumAmount:
          type: number
          format: double
          example: 0
        settlementAmount:
          type: number
          format: double
          example: 0
        payDay:
          type: string
          example: 12-03-2023
        createDate:
          type: string
          example: 12-03-2023
    CreditCardResponseDto:
      required:
        - creditCard
        - statusDto
      type: object
      properties:
        statusDto:
          $ref: '#/components/schemas/MessageStatus'
        creditCard:
          type: array
          items:
            $ref: '#/components/schemas/CreditCard'
    CreditCardRequestDto:
      required:
        - createDate
        - creditCardAccount
        - creditCardId
        - creditLine
        - customerId
        - feeMonth
        - lineAvailable
        - minimumAmount
        - payDay
        - settlementAmount
      type: object
      properties:
        id:
          type: string
          example: 6413619bf92d28ec5a2e104e
        customerId:
          type: string
          example: 6413619bf92d28ec5a2e104e
        creditLine:
          type: number
          format: double
          example: 1000
        creditCardAccount:
          type: string
          example: 6413619bf92d28ec5a2e104e
        creditCardId:
          type: string
          example: 1111-2222-3333-4444
        lineAvailable:
          type: number
          format: double
          example: 1000
        feeMonth:
          type: number
          format: double
          example: 0
        minimumAmount:
          type: number
          format: double
          example: 0
        settlementAmount:
          type: number
          format: double
          example: 0
        payDay:
          type: string
          example: 12-03-2023
        createDate:
          type: string
          example: 12-03-2023
    CreditCardPayRequestDto:
      required:
        - amount
        - customerId
        - id
        - payDay
      type: object
      properties:
        id:
          type: string
          example: 6413619bf92d28ec5a2e104e
        customerId:
          type: string
          example: 6413619bf92d28ec5a2e104e
        amount:
          type: number
          format: double
          example: 500
        payDay:
          type: string
          example: 12-03-2023
    CreditCardPayResponseDto:
      required:
        - statusDto
      type: object
      properties:
        statusDto:
          $ref: '#/components/schemas/MessageStatus'
    CreditCardConsumeRequestDto:
      required:
        - amount
        - customerId
        - id
      type: object
      properties:
        id:
          type: string
          example: 6413619bf92d28ec5a2e104e
        customerId:
          type: string
          example: 6413619bf92d28ec5a2e104e
        amount:
          type: number
          format: double
          example: 500
    CreditCardConsumeResponseDto:
      required:
        - statusDto
      type: object
      properties:
        statusDto:
          $ref: '#/components/schemas/MessageStatus'
    Account:
      required:
        - amount
        - comMaintenance
        - createDate
        - customerId
        - id
        - monthlyMovements
        - movementDate
        - status
        - typeAccount
        - typeCustomer
      type: object
      properties:
        id:
          type: string
          example: 6413619bf92d28ec5a2e104e
        typeAccount:
          type: string
          example: AHORRO
        typeCustomer:
          type: string
          example: PERSONAL
        customerId:
          type: string
          example: 6413619bf92d28ec5a2e104e
        comMaintenance:
          type: number
          format: double
          example: 0
        movementDate:
          type: string
          example: 15-03-2023
        monthlyMovements:
          type: integer
          example: 10
        amount:
          type: number
          format: double
          example: 500
        createDate:
          type: string
          example: 12-03-2023
        status:
          type: integer
          example: 0
    AccountResponseDto:
      required:
        - account
        - statusDto
      type: object
      properties:
        statusDto:
          $ref: '#/components/schemas/MessageStatus'
        account:
          type: array
          items:
            $ref: '#/components/schemas/Account'
    AccountRequestDto:
      required:
        - amount
        - comMaintenance
        - createDate
        - customerId
        - monthlyMovements
        - movementDate
        - status
        - typeAccount
        - typeCustomer
      type: object
      properties:
        id:
          type: string
          example: 6413619bf92d28ec5a2e104e
        typeAccount:
          type: string
          example: AHORRO
        typeCustomer:
          type: string
          example: PERSONAL
        customerId:
          type: string
          example: 6413619bf92d28ec5a2e104e
        comMaintenance:
          type: number
          format: double
          example: 0
        movementDate:
          type: string
          example: 15-03-2023
        monthlyMovements:
          type: integer
          example: 10
        amount:
          type: number
          format: double
          example: 500
        createDate:
          type: string
          example: 12-03-2023
        status:
          type: integer
          example: 0
    AccountDepositResponseDto:
      required:
        - statusDto
      type: object
      properties:
        statusDto:
          $ref: '#/components/schemas/MessageStatus'
    AccountDepositRequestDto:
      required:
        - amount
        - customerId
        - id
        - movementDate
        - typeAccount
        - typeCustomer
      type: object
      properties:
        id:
          type: string
          example: 6413619bf92d28ec5a2e104e
        typeAccount:
          type: string
          example: AHORRO
        typeCustomer:
          type: string
          example: PERSONAL
        customerId:
          type: string
          example: 6413619bf92d28ec5a2e104e
        movementDate:
          type: string
          example: 15-03-2023
        amount:
          type: number
          format: double
          example: 500
    AccountWithdrawResponseDto:
      required:
        - statusDto
      type: object
      properties:
        statusDto:
          $ref: '#/components/schemas/MessageStatus'
    AccountWithdrawRequestDto:
      required:
        - amount
        - customerId
        - id
        - movementDate
        - typeAccount
        - typeCustomer
      type: object
      properties:
        id:
          type: string
          example: 6413619bf92d28ec5a2e104e
        typeAccount:
          type: string
          example: AHORRO
        typeCustomer:
          type: string
          example: PERSONAL
        customerId:
          type: string
          example: 6413619bf92d28ec5a2e104e
        movementDate:
          type: string
          example: 15-03-2023
        amount:
          type: number
          format: double